requirements.txt
albumentations==1.4.1
diffusers==0.20.2
einops==0.7.0
imageio==2.31.1
imageio-ffmpeg==0.4.8
ipython==8.12.3
kornia==0.7.0
matplotlib==3.8.3
natsort==8.4.0
numpy==1.26.4
omegaconf==2.3.0
opencv_python==4.9.0.80
Pillow==10.2.0
PyMCubes==0.1.4
pyrender==0.1.45
pytorch_lightning==2.0.8
PyYAML==6.0.1
Requests==2.31.0
rich==13.7.1
scipy==1.12.0
setuptools==68.2.2
six==1.16.0
taming_transformers_rom1504==0.0.6
torch_ema==0.3
torchmetrics==1.3.1
tqdm==4.66.2
transformers==4.38.2
trimesh==4.2.0
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
gdown==5.1.0

